requirements.txt
requests==2.31.0
jpholiday==0.1.7
python-dotenv==1.0.0
pandas==2.1.4 